Cleaned up the pfcpiface controller.
Separated out BESS specific GRPC calls from pfcpiface.go
Separated out PFCP IE parsing from pfcpiface.go.
