+++ pfcpiface/pfcpiface.go
@@ -4,7 +4,6 @@
 package main
 
 import (
-	"context"
 	"log"
 	"net"
 	"time"
@@ -29,208 +28,6 @@ type sessRecord struct {
 
 var sessions map[uint64]sessRecord
 
-func parsePDRFromPFCPSessEstReqPayload(upf *upf, sereq *message.SessionEstablishmentRequest, fseid *ie.FSEIDFields) (pdrs []pdr, fars []far) {
-
-	pdrList := make([]pdr, 0, MaxItems)
-	farList := make([]far, 0, MaxItems)
-
-	/* read PDR(s) */
-	ies1, err := ie.ParseMultiIEs(sereq.Payload)
-	if err != nil {
-		log.Println("Failed to parse sereq for IEs!")
-		return pdrList, farList
-	}
-	/*
-	 * Iteratively go through all IEs. You can't use ie.CreatePDR or ie.CreateFAR since a single
-	 * message can carry multiple CreatePDR & CreateFAR messages.
-	 */
-	for _, ie1 := range ies1 {
-		switch ie1.Type {
-		case ie.CreatePDR:
-			var srcIface uint8
-			var teid uint32
-			var ueIP4 net.IP
-
-			/* reset outerHeaderRemoval to begin with */
-			outerHeaderRemoval := uint8(0)
-
-			pdrID, err := ie1.PDRID()
-			if err != nil {
-				log.Println("Could not read PDR!")
-				return pdrList, farList
-			}
-
-			pdi, err := ie1.PDI()
-			if err != nil {
-				log.Println("Could not read PDI!")
-				return pdrList, farList
-			}
-
-			_, err = ie1.OuterHeaderRemoval()
-			if err == nil {
-				res, err := ie1.OuterHeaderRemovalDescription()
-				if res == 0 && err == nil { // 0 == GTP-U/UDP/IPv4
-					outerHeaderRemoval = 1
-				}
-			}
-
-			for _, ie2 := range pdi {
-				switch ie2.Type {
-				case ie.SourceInterface:
-					srcIface, err = ie2.SourceInterface()
-					if err != nil {
-						log.Println("Failed to parse Source Interface IE!")
-					} else {
-						if srcIface == ie.SrcInterfaceCPFunction {
-							log.Println("Detected src interface cp function. Ignoring for the time being")
-						}
-					}
-				case ie.FTEID:
-					fteid, err := ie2.FTEID()
-					if err != nil {
-						log.Println("Failed to parse FTEID IE")
-					} else {
-						teid = fteid.TEID
-					}
-				case ie.UEIPAddress:
-					ueip4, err := ie2.UEIPAddress()
-					if err != nil {
-						log.Println("Failed to parse UE IP address")
-					} else {
-						ueIP4 = ueip4.IPv4Address
-					}
-				case ie.SDFFilter:
-					// Do nothing for the time being
-				case ie.QFI:
-					// Do nothing for the time being
-				}
-			}
-
-			farID, err := ie1.FARID()
-			if err != nil {
-				log.Println("Could not read FAR ID!")
-				return pdrList, farList
-			}
-
-			// uplink PDR may not have UE IP address IE
-			// FIXIT/TODO/XXX Move this inside SrcInterfaceAccess IE check??
-			var ueIP uint32
-			var ueIPMask uint32
-			if len(ueIP4) == 0 {
-				ueIP = 0
-				ueIPMask = 0
-			} else {
-				ueIP = ip2int(ueIP4)
-				ueIPMask = 0xFFFFFFFF
-			}
-
-			// populated everything for PDR, and set the right pdr_
-			if srcIface == ie.SrcInterfaceAccess {
-				pdrU := pdr{
-					srcIface:     access,
-					eNBTeid:      teid,
-					dstIP:        ueIP,
-					srcIfaceMask: 0xFF,
-					eNBTeidMask:  0xFFFFFFFF,
-					dstIPMask:    ueIPMask,
-					pdrID:        uint32(pdrID),
-					fseID:        uint32(fseid.SEID), // fseID currently being truncated to uint32 <--- FIXIT/TODO/XXX
-					ctrID:        0,                  // ctrID currently not being set <--- FIXIT/TODO/XXX
-					farID:        uint8(farID),
-					needDecap:    outerHeaderRemoval,
-				}
-				pdrList = append(pdrList, pdrU)
-			} else if srcIface == ie.SrcInterfaceCore {
-				pdrD := pdr{
-					srcIface:     core,
-					srcIP:        ueIP,
-					srcIfaceMask: 0xFF,
-					srcIPMask:    ueIPMask,
-					pdrID:        uint32(pdrID),
-					fseID:        uint32(fseid.SEID), // fseID currently being truncated to uint32 <--- FIXIT/TODO/XXX
-					ctrID:        0,                  // ctrID currently not being set <--- FIXIT/TOOD/XXX
-					farID:        uint8(farID),       // farID currently not being set <--- FIXIT/TODO/XXX
-					needDecap:    outerHeaderRemoval,
-				}
-				pdrList = append(pdrList, pdrD)
-			}
-
-		case ie.CreateFAR:
-			if err != nil {
-				log.Println("Failed to parse FAR IE!")
-			} else {
-				farID, err := ie1.FARID()
-				if err != nil {
-					log.Println("Could not read FAR ID!")
-					return pdrList, farList
-				}
-				/* Read outerheadercreation from payload (if it exists) */
-				var eNBTeid uint32
-				eNBIP := uint32(0)
-				n6IP4 := uint32(0)
-				tunnelType := uint8(0)
-				ies2, err := ie1.ForwardingParameters()
-				if err != nil {
-					log.Println("Unable to find ForwardingParameters!")
-				} else {
-					for _, ie2 := range ies2 {
-						switch ie2.Type {
-						case ie.OuterHeaderCreation:
-							outerheadercreationfields, err := ie2.OuterHeaderCreation()
-							if err != nil {
-								log.Println("Unable to parse OuterHeaderCreationFields!")
-							} else {
-								eNBTeid = outerheadercreationfields.TEID
-								eNBIP = ip2int(outerheadercreationfields.IPv4Address)
-								n6IP4 = ip2int(upf.n6IP)
-								tunnelType = uint8(1)
-							}
-						}
-					}
-				}
-				far := far{
-					farID:       uint8(farID),       // farID currently being truncated to uint8 <--- FIXIT/TODO/XXX
-					fseID:       uint32(fseid.SEID), // fseID currently being truncated to uint32 <--- FIXIT/TODO/XXX
-					action:      farForwardU,
-					tunnelType:  tunnelType,
-					s1uIP:       n6IP4,
-					eNBIP:       eNBIP,
-					eNBTeid:     eNBTeid,
-					UDPGTPUPort: udpGTPUPort,
-				}
-				farList = append(farList, far)
-			}
-
-		}
-	}
-
-	return pdrList, farList
-}
-
-func deleteAllSessions(upf *upf) {
-	/* create context, pause daemon, insert PDR(s), and resume daemon */
-	ctx, cancel := context.WithTimeout(context.Background(), Timeout)
-	defer cancel()
-	done := make(chan bool)
-	calls := 5
-
-	err := upf.pauseAll()
-	if err != nil {
-		log.Fatalln("Unable to pause BESS:", err)
-	}
-	upf.removeAllPDRs(ctx, done)
-	upf.removeAllFARs(ctx, done)
-	upf.removeAllCounters(ctx, done, "preQoSCounter")
-	upf.removeAllCounters(ctx, done, "postDLQoSCounter")
-	upf.removeAllCounters(ctx, done, "postULQoSCounter")
-
-	upf.GRPCJoin(calls, Timeout, done)
-	err = upf.resumeAll()
-	if err != nil {
-		log.Fatalln("Unable to resume BESS:", err)
-	}
-}
-
 func pfcpifaceMainLoop(upf *upf, n3ip string, sourceIP string) {
 	log.Println("pfcpifaceMainLoop@" + upf.fqdnHost + " says hello!!!")
 
@@ -251,6 +48,14 @@ func pfcpifaceMainLoop(upf *upf, n3ip string, sourceIP string) {
 	// flag to check SMF/SPGW-C is connected
 	cpConnected := false
 
+	// cleanup the pipeline
+	cleanupSessions := func() {
+		if cpConnected {
+			sendDeleteAllSessionsMsgtoUPF(upf)
+			cpConnected = false
+		}
+	}
+
 	// Initialize pkt buf
 	buf := make([]byte, PktBufSz)
 	// Initialize sessions map
@@ -259,7 +64,7 @@ func pfcpifaceMainLoop(upf *upf, n3ip string, sourceIP string) {
 	for {
 		err := conn.SetReadDeadline(time.Now().Add(readTimeout))
 		if err != nil {
-			log.Fatalln("Unable to set deadline for read")
+			log.Fatalln("Unable to set deadline for read:", err)
 		}
 		// blocking read
 		n, addr, err := conn.ReadFrom(buf)
@@ -267,10 +72,7 @@ func pfcpifaceMainLoop(upf *upf, n3ip string, sourceIP string) {
 			if err, ok := err.(net.Error); ok && err.Timeout() {
 				// do nothing for the time being
 				log.Println(err)
-				if cpConnected {
-					deleteAllSessions(upf)
-					cpConnected = false
-				}
+				cleanupSessions()
 				continue
 			}
 			log.Fatalln("Read error:", err)
@@ -301,6 +103,7 @@ func pfcpifaceMainLoop(upf *upf, n3ip string, sourceIP string) {
 			outgoingMessage = handleSessionDeletionRequest(upf, msg, addr, sourceIP)
 		case message.MsgTypeAssociationReleaseRequest:
 			outgoingMessage = handleAssociationReleaseRequest(msg, addr, sourceIP, n3ip)
+			cleanupSessions()
 		default:
 			log.Println("Message type: ", msg.MessageTypeName(), " is currently not supported")
 			continue
@@ -396,28 +199,7 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 	/* Read CreatePDRs and CreateFARs from payload */
 	pdrs, fars := parsePDRFromPFCPSessEstReqPayload(upf, sereq, fseid)
 
-	/* create context, pause daemon, insert PDR(s), and resume daemon */
-	ctx, cancel := context.WithTimeout(context.Background(), Timeout)
-	defer cancel()
-	done := make(chan bool)
-	calls := len(pdrs) + len(fars)
-
-	err = upf.pauseAll()
-	if err != nil {
-		log.Fatalln("Unable to pause BESS:", err)
-	}
-	for _, pdr := range pdrs {
-		upf.addPDR(ctx, done, pdr)
-	}
-	for _, far := range fars {
-		upf.addFAR(ctx, done, far)
-	}
-
-	upf.GRPCJoin(calls, Timeout, done)
-	err = upf.resumeAll()
-	if err != nil {
-		log.Fatalln("Unable to resume BESS:", err)
-	}
+	upf.sendMsgToUPF("add", pdrs, fars)
 
 	// Adding current session details to the hash map
 	sessItem := sessRecord{
@@ -458,6 +240,9 @@ func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Ad
 	/* fetch FSEID */
 	fseid := (mySEID(smreq.SEID()))
 
+	/* initialize farList */
+	fars := make([]far, 0, MaxItems)
+
 	/* read FAR(s). These can be multiple */
 	ies1, err := ie.ParseMultiIEs(smreq.Payload)
 	if err != nil {
@@ -470,69 +255,17 @@ func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Ad
 		for _, ie1 := range ies1 {
 			switch ie1.Type {
 			case ie.UpdateFAR:
-				/* check for updatefar, and fetch FAR ID */
-				farID, err := ie1.FARID()
-				if err != nil {
-					log.Println("Unable to find updateFAR's FAR ID!")
-					continue
+				if f := parseUpdateFAR(ie1, fseid, upf.n3IP); f != nil {
+					fars = append(fars, *f)
 				}
-				var eNBTeid uint32
-				eNBIP := uint32(0)
-				n6IP4 := uint32(0)
-				tunnelType := uint8(0)
-				/* Read UpdateFAR from payload */
-				ies2, err := ie1.UpdateForwardingParameters()
-				if err != nil {
-					log.Println("Unable to find UpdateForwardingParameters!")
-				} else {
-					for _, ie2 := range ies2 {
-						switch ie2.Type {
-						case ie.OuterHeaderCreation:
-							outerheadercreationfields, err := ie2.OuterHeaderCreation()
-							if err != nil {
-								log.Println("Unable to parse OuterHeaderCreationFields!")
-							} else {
-								eNBTeid = outerheadercreationfields.TEID
-								eNBIP = ip2int(outerheadercreationfields.IPv4Address)
-								n6IP4 = ip2int(upf.n3IP)
-								tunnelType = uint8(1)
-							}
-						case ie.DestinationInterface:
-							// Do nothing for the time being
-						}
-					}
-				}
-				far := far{
-					farID:       uint8(farID),  // farID currently being truncated to uint8 <--- FIXIT/TODO/XXX
-					fseID:       uint32(fseid), // fseID currently being truncated to uint32 <--- FIXIT/TODO/XXX
-					action:      farForwardD,
-					tunnelType:  tunnelType,
-					s1uIP:       n6IP4,
-					eNBIP:       eNBIP,
-					eNBTeid:     eNBTeid,
-					UDPGTPUPort: udpGTPUPort,
-				}
-				/* create context */
-				ctx, cancel := context.WithTimeout(context.Background(), Timeout)
-				defer cancel()
-				done := make(chan bool)
-				calls := 1
-
-				// pause daemon, and then insert FAR, finally resume
-				err = upf.pauseAll()
-				if err != nil {
-					log.Fatalln("Unable to pause BESS:", err)
-				}
-				upf.addFAR(ctx, done, far)
-
-				upf.GRPCJoin(calls, Timeout, done)
-				err = upf.resumeAll()
-				if err != nil {
-					log.Fatalln("Unable to resume BESS:", err)
-				}
+			default:
+				/* more will be added later */
 			}
 		}
 	}
+
+	upf.sendMsgToUPF("add", nil, fars)
+
 	// Build response message
 	smres, err := message.NewSessionModificationResponse(0, /* MO?? <-- what's this */
 		0,                                    /* FO <-- what's this? */
@@ -587,30 +320,7 @@ func handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, 
 	/* retrieve sessionRecord */
 	sessItem := sessions[mySEID(sdreq.SEID())]
 
-	/* create context */
-	ctx, cancel := context.WithTimeout(context.Background(), 1000*time.Millisecond)
-	defer cancel()
-
-	done := make(chan bool)
-	calls := len(sessItem.pdrs) + len(sessItem.fars)
-
-	// pause daemon, and then delete FAR(s) and PDR(s), finally resume
-	err := upf.pauseAll()
-	if err != nil {
-		log.Fatalln("Unable to pause BESS:", err)
-	}
-	for _, pdr := range sessItem.pdrs {
-		upf.delPDR(ctx, done, pdr)
-	}
-	for _, far := range sessItem.fars {
-		upf.delFAR(ctx, done, far)
-	}
-
-	upf.GRPCJoin(calls, Timeout, done)
-	err = upf.resumeAll()
-	if err != nil {
-		log.Fatalln("Unable to resume BESS:", err)
-	}
+	upf.sendMsgToUPF("del", sessItem.pdrs, sessItem.fars)
 
 	/* delete sessionRecord */
 	delete(sessions, mySEID(sdreq.SEID()))
